backend/requirements.txt
flask
flask-cors
mysql-connector-python
werkzeug
python-dotenv
faiss-cpu
python-docx
google-generativeai
langchain-text-splitters
easyocr
PyMuPDF
opencv-python
numpy

# LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
typing-extensions>=4.0.0
pydantic>=2.0.0
